move files into app_command_parser folder
