+++ components/suggestion/command_provider/app_command_parser/app_command_parser.test.ts
